Fix nonzero/NA disagreement, restore radius validation (adversarial review)
Two HIGH findings survived the previous commit, plus two lower ones and a
pre-existing crash found in the same area. All reproduced before fixing.

  * `nonzero = TRUE` disagreed about missing values. The compiled filter dropped
    NA/NaN voxels; the R filter was `vals != 0`, which evaluates to NA for a
    missing voxel, and an NA logical subscript emits an all-NA row -- so a
    ROIVolWindow could come back with NA NA NA coordinates. On a 3x3x3 volume
    with one NA voxel, eager and lazy searchlights differed on 21 of 25 windows,
    and both differed from spherical_roi(). The rule now exists in exactly one
    place: the compiled core takes the flattened volume and filters on
    "present and non-zero", and the R builders match it. This also removes the
    separate `keep` buffer, halving what a searchlight plan retains.

  * The lazy iterators had stopped validating `radius < min(spacing)`, which
    make_spherical_grid() enforces for the single-ROI builders, so
    searchlight(eager = FALSE) and searchlight_coords() silently returned
    degenerate one-voxel windows where every other path aborts.

  * sphere_roi_at_cpp() narrowed parent_index with a C cast; past 2^31 voxels
    that is implementation-defined. It now reports NA_INTEGER, matching the R
    builders' overflow behaviour.

  * The iterator closures captured `mask`, `mask.idx`, `cores` and friends they
    never used. Built with local() now, they retain only the plan and the centre
    grid: 9.25 MB -> 5.85 MB on an 80x80x60 mask.

Also fixed, pre-existing and out of the reviewed diff but a hard crash in an
exported function: index_to_grid(NeuroSpace(c(65536L, 65536L, 2L)), 1:3) killed
the session with SIGFPE, because the stride accumulator in indexToGridCpp was
`int` and wrapped to zero before a division. It is now 64-bit.

The review also refuted the constructor's "~3 us" comment: that is the mechanism,
not the function, which measures ~11 us once its argument checks are included
(still 68x below new()). The comment is corrected. Two guards were tightened at
the same time: `.Data` is now required to be double or integer, since is.vector()
is TRUE for lists and as.vector() is an internal generic whose S3 methods may
return anything; and the slot-type predicates reject classed values, which
is.integer()/is.matrix() accept but the slot definitions do not.

Verified: golden sweep extended with an NA/NaN volume section (8665 comparisons,
0 mismatches); suite 3737 pass / 0 fail / unchanged 25 shim errors; 1470
assertions in test-roi-series-fastpaths.R, including a test that asserts the
constructor's hand-written invariants and the class validity function agree, so
the duplication cannot drift silently.

+++ src/indexFuns.cpp
@@ -79,11 +79,14 @@ namespace indexfuns {
             int tmp = 1 + wh1 % array_dim(0);
             IntegerVector wh = IntegerVector(rank, tmp);
             if (rank >= 2) {
-                int denom = 1;
+                // int64: the running product of dimensions overflows int for
+                // volumes past ~2^31 voxels, and once it wraps to 0 the division
+                // below raises SIGFPE and kills the session.
+                int64_t denom = 1;
                 for (int j = 1; j < rank; j++) {
-                    denom = denom * array_dim(j-1);
-                    int nextd1 = (int)wh1/denom;
-                    wh(j) = 1 + nextd1 % array_dim(j);
+                    denom = denom * (int64_t) array_dim(j-1);
+                    int64_t nextd1 = (int64_t) wh1 / denom;
+                    wh(j) = (int)(1 + nextd1 % (int64_t) array_dim(j));
                 }
             }
             omat.row(i) = wh;
